task-#68: Preventing potential timing based side channel attack
On SRP6 implementation, when validating M1 and M2, the computing time
must be the same independently whether the evidence is valid or not.
This must be in this way, to prevent timing based side channel attacks
which might take advantage of private information leaked if that
validation time is not the same on both validation paths.

Prepared to released version 1.1.4