task-#71: Easy to use AES cipher
Implemented an easy to use AES cipher which makes easy to encrypt and decrypt
using AES algorithm, hiding the underlying complexities, and requiring only the content
to encrypt/decrypt and the secret key, regardless of the block mode of operation used.

In case the implemented block mode of operation is IV based, then the implementation will
generate the IV on the fly and will prefix/read it to/from the output/input.